Completed Sprint Challenge.
