+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_list_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
