+++ app/src/main/java/com/lambdaschool/shoppinglist/model/ShoppingRepository.kt
@@ -1,4 +1,4 @@
-package com.lambdaschool.shoppinglist
+package com.lambdaschool.shoppinglist.model
 
 import com.lambdaschool.sprint2_challenge.itemIds
 import com.lambdaschool.sprint2_challenge.itemNames
@@ -8,7 +8,13 @@ class ShoppingRepository {
         val shoppingList = mutableListOf<Shopping>()
         fun createShoppingList() {
             for (i in 0 until itemNames.size) {
-                shoppingList.add(Shopping(itemNames[i], itemIds[i], isAdded = false))
+                shoppingList.add(
+                    Shopping(
+                        itemNames[i],
+                        itemIds[i],
+                        isAdded = false
+                    )
+                )
             }
         }
     }
