+++ app/src/main/res/layout/activity_main.xml
@@ -5,14 +5,15 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
 
     <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_item_list"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="@+id/btn_send_list"
-            app:layout_constraintTop_toTopOf="parent">
+        android:id="@+id/rv_item_list"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintBottom_toTopOf="@+id/btn_send_list"
+        app:layout_constraintTop_toTopOf="parent">
+
     </androidx.recyclerview.widget.RecyclerView>
 
     <Button
